refactoring: move the npm package installation code from the consumer to a new file npm-client/install-packages.js
